Group UI methods to allow overriding.
Also, in Javascript, provide the target element into which to render the UI.
