+++ src/diagonal.works/b6/cmd/b6/js/main.js
@@ -1,3 +1,3 @@
 import main from "./b6.js";
 
-main();
+main("#b6");
