+++ src/diagonal.works/b6/cmd/b6/js/static/b6.css
@@ -63,7 +63,7 @@ a:hover {
     text-decoration: underline;
 }
 
-#map {
+.map {
     position: absolute;
     top: 0px;
     left: 0px;
@@ -119,7 +119,7 @@ a:hover {
     cursor: pointer;
 }
 
-#shell {
+.shell {
     position: absolute;
     left: 10px;
     top: 42px;
@@ -134,7 +134,7 @@ a:hover {
     border-radius: 3px;
 }
 
-#shell.closed {
+.shell.closed {
     left: -10010px;
 }
 
@@ -142,17 +142,17 @@ a:hover {
     cursor: pointer;
 }
 
-form.shell {
+form.shell-form {
     display: flex;
     flex-direction: row;
 }
 
-form.shell > .prompt {
+form.shell-form > .prompt {
     color: #528dff;
     margin-right: 8px;
 }
 
-form.shell > input {
+form.shell-form > input {
     border: 0px;
     padding: 0px;
     outline: 0px;
@@ -707,41 +707,41 @@ form.shell > input {
     border-bottom-right-radius: 4px;
 }
 
-#dock {
+.dock {
     position: absolute;
     top: 100px;
     left: 10px;
 }
 
-#dock > .stack {
+.dock > .stack {
     margin-bottom: 6px;
 }
 
-#dock > .stack.closed > .substack {
+.dock > .stack.closed > .substack {
     display: none;
 }
 
-#dock > .stack.closed > .substack:first-child {
+.dock > .stack.closed > .substack:first-child {
     display: block;
 }
 
-#dock > .stack.closed > .substack:first-child > .scrollable > .line:last-child {
+.dock > .stack.closed > .substack:first-child > .scrollable > .line:last-child {
     border-bottom-left-radius: 4px;
     border-bottom-right-radius: 4px;
 }
 
-#dock > .stack.closed {
+.dock > .stack.closed {
     border: 1px solid #dbdeec; /* Graphite 30 */
 }
 
-#dock > .stack.closed .line-question {
+.dock > .stack.closed .line-question {
     color: #4f5a7d; /* Graphite 80 */
 }
 
-#dock > .stack.closed .line-question:hover {
+.dock > .stack.closed .line-question:hover {
     background: #f8f9fc /* Graphite 10 */
 }
 
-#dock > .stack > .substack > .scrollable {
+.dock > .stack > .substack > .scrollable {
     max-height: 282px;
 }
