+++ src/diagonal.works/b6/cmd/b6/js/static/index.html
@@ -8,9 +8,7 @@
         <link rel="icon" type="image/png" sizes="16x16" href="/images/favicon-16x16.png"/>
     </head>
     <body>
-        <div id="map"></div>
-        <div id="shell" class="closed"></div>
-        <div id="dock"></div>
+        <div id="b6"></div>
     </body>
     <script src="/bundle.js"></script>
 </html>
