+++ src/diagonal.works/b6/ui/lines.go
@@ -1,47 +1,16 @@
 package ui
 
 import (
-	"context"
-	"encoding/json"
 	"fmt"
-	"io"
-	"log"
-	"net/http"
 	"reflect"
 	"strconv"
 	"strings"
 
 	"diagonal.works/b6"
 	"diagonal.works/b6/api"
-	"diagonal.works/b6/api/functions"
-	"diagonal.works/b6/geojson"
-	"diagonal.works/b6/ingest"
 	pb "diagonal.works/b6/proto"
-	"diagonal.works/b6/renderer"
 	"github.com/golang/geo/s2"
-	"google.golang.org/protobuf/encoding/protojson"
 )
-
-type DefaultUIRenderer struct {
-	BasemapRules    renderer.RenderRules
-	FunctionSymbols api.FunctionSymbols
-	World           b6.World
-}
-
-func (d *DefaultUIRenderer) Render(response *UIResponseJSON, value interface{}, context b6.CollectionFeature, locked bool) error {
-	if err := d.fillResponseFromResult(response, value); err == nil {
-		shell := &pb.ShellLineProto{
-			Functions: make([]string, 0),
-		}
-		shell.Functions = fillMatchingFunctionSymbols(shell.Functions, value, d.FunctionSymbols)
-		response.Proto.Stack.Substacks = append(response.Proto.Stack.Substacks, &pb.SubstackProto{
-			Lines: []*pb.LineProto{{Line: &pb.LineProto_Shell{Shell: shell}}},
-		})
-		return nil
-	} else {
-		return d.fillResponseFromResult(response, err)
-	}
-}
 
 func getStringExpression(f b6.Feature, key string) *pb.NodeProto {
 	return &pb.NodeProto{
@@ -135,160 +104,6 @@ func fillMatchingFunctionSymbols(symbols []string, result interface{}, functions
 		}
 	}
 	return symbols
-}
-
-func NewUIHandler(renderer UIRenderer, w ingest.MutableWorld, options api.Options) *UIHandler {
-	return &UIHandler{
-		Renderer:        renderer,
-		World:           w,
-		Options:         options,
-		FunctionSymbols: functions.Functions(),
-		Adaptors:        functions.Adaptors(),
-	}
-}
-
-type UIResponseProtoJSON pb.UIResponseProto
-
-func (b *UIResponseProtoJSON) MarshalJSON() ([]byte, error) {
-	return protojson.Marshal((*pb.UIResponseProto)(b))
-}
-
-func (b *UIResponseProtoJSON) UnmarshalJSON(buffer []byte) error {
-	return protojson.Unmarshal(buffer, (*pb.UIResponseProto)(b))
-}
-
-type UIResponseJSON struct {
-	Proto   *UIResponseProtoJSON `json:"proto,omitempty"`
-	GeoJSON []geojson.GeoJSON    `json:"geoJSON,omitempty"`
-}
-
-func NewUIResponseJSON() *UIResponseJSON {
-	return &UIResponseJSON{
-		Proto: &UIResponseProtoJSON{
-			Stack: &pb.StackProto{},
-		},
-	}
-}
-
-func (u *UIResponseJSON) AddGeoJSON(g geojson.GeoJSON) {
-	u.GeoJSON = append(u.GeoJSON, g)
-	u.Proto.GeoJSON = append(u.Proto.GeoJSON, &pb.GeoJSONProto{
-		Index: int32(len(u.GeoJSON) - 1),
-	})
-}
-
-type UIHandler struct {
-	World           ingest.MutableWorld
-	Renderer        UIRenderer
-	Options         api.Options
-	FunctionSymbols api.FunctionSymbols
-	Adaptors        api.Adaptors
-}
-
-func (u *UIHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	request := &pb.UIRequestProto{}
-	response := NewUIResponseJSON()
-
-	if r.Method == "GET" {
-		request.Expression = r.URL.Query().Get("e")
-	} else if r.Method == "POST" {
-		var err error
-		var body []byte
-		if body, err = io.ReadAll(r.Body); err == nil {
-			r.Body.Close()
-			err = protojson.Unmarshal(body, request)
-		}
-		if err != nil {
-			log.Printf("Bad request body")
-			http.Error(w, "Bad request body", http.StatusBadRequest)
-			return
-		}
-	} else {
-		log.Printf("Bad method")
-		http.Error(w, "Bad method", http.StatusMethodNotAllowed)
-		return
-	}
-
-	if request.Expression == "" && request.Node == nil {
-		log.Printf("No expression")
-		http.Error(w, "No expression", http.StatusBadRequest)
-		return
-	}
-
-	var root b6.CollectionFeature
-	if request.Context != nil && request.Context.Type == pb.FeatureType_FeatureTypeCollection {
-		root = b6.FindCollectionByID(b6.NewFeatureIDFromProto(request.Context).ToCollectionID(), u.World)
-	}
-
-	var expression b6.Expression
-	if request.Expression != "" {
-		var err error
-		if request.Node == nil {
-			expression, err = api.ParseExpression(request.Expression)
-		} else {
-			var lhs b6.Expression
-			if err = lhs.FromProto(request.Node); err == nil {
-				expression, err = api.ParseExpressionWithLHS(request.Expression, lhs)
-			}
-		}
-		if err != nil {
-			u.Renderer.Render(response, err, root, request.Locked)
-			var substack pb.SubstackProto
-			fillSubstackFromError(&substack, err)
-			response.Proto.Stack.Substacks = append(response.Proto.Stack.Substacks, &substack)
-			sendUIResponse(response, w)
-			return
-		}
-	} else {
-		expression.FromProto(request.Node)
-	}
-	expression = api.Simplify(expression, u.FunctionSymbols)
-
-	if !request.Locked {
-		substack := &pb.SubstackProto{}
-		fillSubstackFromExpression(substack, expression, true)
-		if len(substack.Lines) > 0 {
-			response.Proto.Stack.Substacks = append(response.Proto.Stack.Substacks, substack)
-		}
-	}
-
-	if unparsed, ok := api.UnparseExpression(expression); ok {
-		response.Proto.Expression = unparsed
-	}
-
-	var err error
-	if response.Proto.Node, err = expression.ToProto(); err != nil {
-		u.Renderer.Render(response, err, root, request.Locked)
-		sendUIResponse(response, w)
-		return
-	}
-
-	vmContext := api.Context{
-		World:           u.World,
-		FunctionSymbols: u.FunctionSymbols,
-		Adaptors:        u.Adaptors,
-		Context:         context.Background(),
-	}
-	vmContext.FillFromOptions(&u.Options)
-
-	result, err := api.Evaluate(expression, &vmContext)
-	if err == nil {
-		err = u.Renderer.Render(response, result, root, request.Locked)
-	}
-	if err != nil {
-		u.Renderer.Render(response, err, root, request.Locked)
-	}
-	sendUIResponse(response, w)
-}
-
-func sendUIResponse(response *UIResponseJSON, w http.ResponseWriter) {
-	output, err := json.Marshal(response)
-	if err != nil {
-		http.Error(w, "Failed to encode response", http.StatusInternalServerError)
-		return
-	}
-	w.Header().Set("Content-Type", "application/json")
-	w.Write([]byte(output))
 }
 
 func highlightInResponse(response *pb.UIResponseProto, id b6.FeatureID) {
@@ -617,155 +432,4 @@ func fillSubstacksFromFeature(substacks []*pb.SubstackProto, f b6.Feature, w b6.
 		substacks = append(substacks, substack)
 	}
 	return substacks
-}
-
-func (d *DefaultUIRenderer) fillResponseFromResult(response *UIResponseJSON, result interface{}) error {
-	p := (*pb.UIResponseProto)(response.Proto)
-	switch r := result.(type) {
-	case error:
-		var substack pb.SubstackProto
-		fillSubstackFromError(&substack, r)
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-	case string:
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, AtomFromString(r))
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-	case b6.ExpressionFeature:
-		// This is not perfect, as it makes original expression that
-		// returned the ExpressionFeature, and the expression from the
-		// feature itself, look like part of the same stack.
-		// TODO: improve the UX for expression features
-		substack := &pb.SubstackProto{}
-		expression := api.AddPipelines(api.Simplify(b6.NewCallExpression(r.Expression(), []b6.Expression{}), d.FunctionSymbols))
-		fillSubstackFromExpression(substack, expression, true)
-		if len(substack.Lines) > 0 {
-			response.Proto.Stack.Substacks = append(response.Proto.Stack.Substacks, substack)
-		}
-		id := b6.MakeCollectionID(r.ExpressionID().Namespace, r.ExpressionID().Value)
-		if c := b6.FindCollectionByID(id, d.World); c != nil {
-			substack := &pb.SubstackProto{}
-			if err := fillSubstackFromCollection(substack, c, p, d.World); err == nil {
-				p.Stack.Substacks = append(p.Stack.Substacks, substack)
-			} else {
-				return err
-			}
-		}
-	case b6.Feature:
-		p.Stack.Substacks = fillSubstacksFromFeature(p.Stack.Substacks, r, d.World)
-		highlightInResponse(p, r.FeatureID())
-	case b6.Query:
-		if q, ok := api.UnparseQuery(r); ok {
-			var substack pb.SubstackProto
-			fillSubstackFromAtom(&substack, AtomFromString(q))
-			p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		} else {
-			// TODO: Improve the rendering of queries
-			var substack pb.SubstackProto
-			fillSubstackFromAtom(&substack, AtomFromString("Query"))
-			p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		}
-	case b6.Tag:
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, AtomFromValue(r, d.World))
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		if !d.BasemapRules.IsRendered(r) {
-			if q, ok := api.UnparseQuery(b6.Tagged(r)); ok {
-				before := pb.MapLayerPosition_MapLayerPositionEnd
-				if r.Key == "#boundary" {
-					before = pb.MapLayerPosition_MapLayerPositionBuildings
-				}
-				p.Layers = append(p.Layers, &pb.MapLayerProto{Query: q, Before: before})
-			}
-		}
-	case *api.HistogramCollection:
-		if err := fillResponseFromHistogram(p, r, d.World); err != nil {
-			return err
-		}
-	case b6.UntypedCollection:
-		substack := &pb.SubstackProto{}
-		if err := fillSubstackFromCollection(substack, r, p, d.World); err == nil {
-			p.Stack.Substacks = append(p.Stack.Substacks, substack)
-		} else {
-			return err
-		}
-	case b6.Area:
-		dimension := 0.0
-		for i := 0; i < r.Len(); i++ {
-			dimension += b6.AreaToMeters2(r.Polygon(i).Area())
-		}
-		atom := &pb.AtomProto{
-			Atom: &pb.AtomProto_Download{
-				Download: fmt.Sprintf("%.2fm² area", dimension),
-			},
-		}
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, atom)
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		response.AddGeoJSON(r.ToGeoJSON())
-	case b6.Path:
-		dimension := b6.AngleToMeters(r.Polyline().Length())
-		atom := &pb.AtomProto{
-			Atom: &pb.AtomProto_Download{
-				Download: fmt.Sprintf("%.2fm path", dimension),
-			},
-		}
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, atom)
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		response.AddGeoJSON(r.ToGeoJSON())
-	case *geojson.FeatureCollection:
-		var label string
-		if n := len(r.Features); n == 1 {
-			label = "1 GeoJSON feature"
-		} else {
-			label = fmt.Sprintf("%d GeoJSON features", n)
-		}
-		atom := &pb.AtomProto{
-			Atom: &pb.AtomProto_Download{
-				Download: fmt.Sprintf(label),
-			},
-		}
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, atom)
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		response.AddGeoJSON(r)
-	case *geojson.Feature:
-		atom := &pb.AtomProto{
-			Atom: &pb.AtomProto_Download{
-				Download: "GeoJSON feature",
-			},
-		}
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, atom)
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		response.AddGeoJSON(r)
-	case *geojson.Geometry:
-		atom := &pb.AtomProto{
-			Atom: &pb.AtomProto_Download{
-				Download: "GeoJSON geometry",
-			},
-		}
-		var substack pb.SubstackProto
-		fillSubstackFromAtom(&substack, atom)
-		p.Stack.Substacks = append(p.Stack.Substacks, &substack)
-		response.AddGeoJSON(geojson.NewFeatureWithGeometry(*r))
-	default:
-		substack := &pb.SubstackProto{
-			Lines: []*pb.LineProto{{
-				Line: &pb.LineProto_Value{
-					Value: &pb.ValueLineProto{
-						Atom: AtomFromValue(r, d.World),
-					},
-				},
-			}},
-		}
-		p.Stack.Substacks = append(p.Stack.Substacks, substack)
-	}
-	switch r := result.(type) {
-	case b6.Geometry:
-		response.Proto.MapCenter = b6.NewPointProtoFromS2Point(b6.Centroid(r))
-	case geojson.GeoJSON:
-		response.Proto.MapCenter = b6.NewPointProtoFromS2Point(r.Centroid().ToS2Point())
-	}
-	return nil
 }
